requirements.txt
pytesseract
opencv-python
pillow
pdf2image
tk